:truck: rename: RegionModal 공통 컴포넌트로 파일 이동
